Necessary refactor to get Flask-Caching to work
The code had to be refactored in order to get Flask-Caching to work for both production and development contexts.

For Flask-Caching to work, the cache initialization and the `server` object instantiation seems to need to be in the same file, and both seem to need to be unnested top-level code as well.